Core/requirements.txt
click>=8.0.0
rich>=10.0.0
pyyaml>=6.0.0
python-dateutil>=2.8.0
pathlib>=1.0.1 